requirements.txt
Django==1.10.6
djangorestframework==3.6.2
django-rest-swagger==2.1.2
pymongo==3.4.0
MongoRouter==0.0.2
wheel==0.24.0